requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==5.0
djangorestframework==3.14.0
drf_yasg==1.21.7
